@@ -35,9 +35,15 @@ class MaterialTransformer implements DataTransformerInterface
                 $materialData['editor_switch'] = $materialItem->getEtherpadEditor() > 0;
 
                 if ($materialItem->getBibKind() != 'none') {
-                    $materialData['biblio_select'] = 'Biblio'.ucfirst($materialItem->getBibKind()).'Type';
+                    $bibKind = $materialItem->getBibKind();
+                    // Bugfix: add backwards compatibility for entries from databases migrated from CommSy < 9
+                    if ($bibKind == 'document' || $bibKind == 'docmanagement') {
+                        $bibKind = 'DocManagement';
+                    }
+                    $materialData['biblio_select'] = 'Biblio'.ucfirst($bibKind).'Type';
                 }
 
+                /** @var \cs_material_item $materialItem */
                 $materialData['biblio_sub']['author'] = $materialItem->getAuthor();
                 $materialData['biblio_sub']['publishing_date'] = $materialItem->getPublishingDate();
                 $materialData['biblio_sub']['common'] = $materialItem->getBibliographicValues();
@@ -61,6 +67,10 @@ class MaterialTransformer implements DataTransformerInterface
                 $materialData['biblio_sub']['foto_copyright'] = $materialItem->getFotoCopyright();
                 $materialData['biblio_sub']['foto_reason'] = $materialItem->getFotoReason();
                 $materialData['biblio_sub']['foto_date'] = $materialItem->getFotoDate();
+                $materialData['biblio_sub']['document_editor'] = $materialItem->getDocumentEditor();
+                $materialData['biblio_sub']['document_maintainer'] = $materialItem->getDocumentMaintainer();
+                $materialData['biblio_sub']['document_release_number'] = $materialItem->getDocumentReleaseNumber();
+                $materialData['biblio_sub']['document_release_date'] = $materialItem->getDocumentReleaseDate();
 
                 $materialData['sections'] = array();
                 foreach($materialItem->getSectionList()->to_array() as $id => $item){
@@ -202,30 +212,36 @@ class MaterialTransformer implements DataTransformerInterface
     }
 
     private function setBibliographic($form_data, $item) {
-        $bibFields = array( 'author',
-                            'publishing_date',
-                            'common',
-                            'bib_kind',
-                            'publisher',
-                            'address',
-                            'edition',
-                            'series',
-                            'volume',
-                            'isbn',
-                            'url',
-                            'url_date',
-                            'editor',
-                            'booktitle',
-                            'issn',
-                            'pages',
-                            'journal',
-                            'issue',
-                            'thesis_kind',
-                            'university',
-                            'faculty',
-                            'foto_copyright',
-                            'foto_reason',
-                            'foto_date');
+        $bibFields = [
+            'author',
+            'publishing_date',
+            'common',
+            'bib_kind',
+            'publisher',
+            'address',
+            'edition',
+            'series',
+            'volume',
+            'isbn',
+            'url',
+            'url_date',
+            'editor',
+            'booktitle',
+            'issn',
+            'pages',
+            'journal',
+            'issue',
+            'thesis_kind',
+            'university',
+            'faculty',
+            'foto_copyright',
+            'foto_reason',
+            'foto_date',
+            'document_editor',
+            'document_maintainer',
+            'document_release_number',
+            'document_release_date'
+]       ;
 
         $converter = $this->legacyEnvironment->getTextConverter();
 
@@ -238,7 +254,7 @@ class MaterialTransformer implements DataTransformerInterface
         }
 
         isset($form_data['value']) ? $form_data['value'] : '';
-        $config = array(
+        $config = [
             array(  'get'       => 'getAuthor',
                     'set'       => 'setAuthor',
                     'value'     => $form_data['author']),
@@ -310,8 +326,20 @@ class MaterialTransformer implements DataTransformerInterface
                     'value'     => $form_data['foto_reason']),
             array(  'get'       => 'getFotoDate',
                     'set'       => 'setFotoDate',
-                    'value'     => $form_data['foto_date'])
-        );
+                    'value'     => $form_data['foto_date']),
+            array(  'get'       => 'getDocumentEditor',
+                    'set'       => 'setDocumentEditor',
+                    'value'     => $form_data['document_editor']),
+            array(  'get'       => 'getDocumentMaintainer',
+                    'set'       => 'setDocumentMaintainer',
+                    'value'     => $form_data['document_maintainer']),
+            array(  'get'       => 'getDocumentReleaseNumber',
+                    'set'       => 'setDocumentReleaseNumber',
+                    'value'     => $form_data['document_release_number']),
+            array(  'get'       => 'getDocumentReleaseDate',
+                    'set'       => 'setDocumentReleaseDate',
+                    'value'     => $form_data['document_release_date']),
+        ];
 
         foreach($config as $method => $detail) {
             if($detail['value'] != call_user_func_array(array($item, $detail['get']), array())) {
