@@ -1,8 +1,8 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Application\Migrations;
 
-use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
 
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
@@ -11,7 +11,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 /**
  * Auto-generated Migration: Please modify to your needs!
  */
-class Version20150623135455 extends AbstractMigration implements ContainerAwareInterface
+final class Version20150623135455 extends AbstractMigration implements ContainerAwareInterface
 {
     private $container;
 
@@ -23,7 +23,7 @@ class Version20150623135455 extends AbstractMigration implements ContainerAwareI
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema) : void
     {
         // this up() migration is auto-generated, please modify it to your needs
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on \'mysql\'.');
@@ -115,7 +115,7 @@ class Version20150623135455 extends AbstractMigration implements ContainerAwareI
         }
     }
 
-    public function postUp(Schema $schema)
+    public function postUp(Schema $schema) : void
     {
         $this->write('updating room configuration in room');
         $this->updateRoomConfiguration("room");
@@ -127,7 +127,7 @@ class Version20150623135455 extends AbstractMigration implements ContainerAwareI
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema) : void
     {
         $this->throwIrreversibleMigrationException();
     }
