@@ -1,8 +1,8 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Application\Migrations;
 
-use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
 
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
@@ -11,7 +11,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 /**
  * Auto-generated Migration: Please modify to your needs!
  */
-class Version20170616103508 extends AbstractMigration implements ContainerAwareInterface
+final class Version20170616103508 extends AbstractMigration implements ContainerAwareInterface
 {
     private $container;
 
@@ -23,7 +23,7 @@ class Version20170616103508 extends AbstractMigration implements ContainerAwareI
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema) : void
     {
         // this up() migration is auto-generated, please modify it to your needs
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on \'mysql\'.');
@@ -77,7 +77,7 @@ class Version20170616103508 extends AbstractMigration implements ContainerAwareI
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema) : void
     {
         // this down() migration is auto-generated, please modify it to your needs
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on \'mysql\'.');
@@ -107,7 +107,8 @@ class Version20170616103508 extends AbstractMigration implements ContainerAwareI
         ');
     }
 
-    public function postUp(Schema $schema) {
+    public function postUp(Schema $schema) : void
+    {
         $this->migrateColorsToCalendars('room');
 
         $this->migrateColorsToCalendars('zzz_room');
