@@ -968,9 +968,12 @@ if(isset($c_smarty) && $c_smarty === true) {
             include_once('pages/picture_getfile.php');
         } elseif ( $environment->inPortal() ) {
             $context_item = $environment->getCurrentContextItem();
-            $filename = 'external_pages/'.$context_item->getItemID().'/cs_external_page_portal_view.php';
-            if ( file_exists($filename) ) {
-                include_once($filename);
+            $externalStylesFilename = 'external_pages/'.$context_item->getItemID().'/cs_external_page_portal_view.php';
+
+            if ($environment->getCurrentModule() !== 'configuration'
+                && $environment->getCurrentModule() !== 'account'
+                && file_exists($externalStylesFilename)) {
+                include_once($externalStylesFilename);
                 $params = array();
                 $params['environment'] = $environment;
                 $params['with_modifying_actions'] = $with_modifying_actions;
